cache storage class is only set once
