+++ test/unit/cacheStorage.test.ts
@@ -1,9 +1,13 @@
 import { keys } from "../../src/container/keys";
 import { CacheStorage } from "../../src/decorators/cache/cacheStorage";
-import { itAddsClassToContainer } from "./utils";
+import {
+  itAddsClassToContainer,
+  itAddsClassToContainerOnlyOnce,
+} from "./utils";
 
 describe("@CacheStorage", () => {
   const name = "CacheStorage";
 
   itAddsClassToContainer(name, CacheStorage, keys.cacheStorageClass);
+  itAddsClassToContainerOnlyOnce(name, CacheStorage);
 });
